# doc/api.yaml
...
openapi: 3.0.4
info:
  title: WASAText API
  description: Professional API for the functionalities of the WASAText messaging app
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            description: Error response for not found resources
            properties:
              error:
                type: string
                description: Error message
                example: Not found
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            description: Error response for unauthorized access
            properties:
              error:
                type: string
                description: Error message
                example: Unauthorized
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            type: object
            description: Error response for invalid input data
            properties:
              error:
                type: string
                description: Error message
                example: Invalid input
  schemas:
    User:
      type: object
      description: A user of the WASAText app
      required: [id, username]
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username for the user
          example: johndoe
        displayName:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*$'
          description: Display name of the user
          example: John Doe
        profilePicture:
          type: string
          format: uri
          minLength: 10
          maxLength: 2048
          pattern: '^https?://.*$'
          description: URL of the user's profile picture
          example: https://example.com/avatar.jpg
...
    Conversation:
      type: object
      description: A conversation (either a chat 1:1 or a group, if is_group is true)
      required: [id, is_group, members]
      properties:
        id:
          type: integer
          description: Unique conversation identifier
          example: 123
        name:
          type: string
          description: Name of the conversation (null for 1:1, set for groups)
          example: Friends Group
          nullable: true
        photo:
          type: string
          format: uri
          description: URL of the group photo (null for 1:1, set for groups)
          example: https://example.com/group.jpg
          nullable: true
        is_group:
          type: boolean
          description: True if this conversation is a group, false if it's a 1:1 chat
          example: true
        members:
          type: array
          minItems: 2
          maxItems: 100
          description: List of conversation members
          items:
            $ref: '#/components/schemas/User'
paths:
...
  /groups:
    post:
      summary: Create group
      description: Create a new group with a name and members
      operationId: addToGroup
      tags: [group]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for creating a group
        content:
          application/json:
            schema:
              type: object
              description: Request body for creating a group
              required: [name, members]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: '^.*$'
                  description: Name of the group
                members:
                  type: array
                  minItems: 1
                  maxItems: 100
                  description: List of usernames to add to the group
                  items:
                    type: string
                    minLength: 3
                    maxLength: 32
                    pattern: '^[a-zA-Z0-9_]+$'
                    description: Username to add to the group
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      summary: List all groups (debug)
      description: Retrieve the list of all groups (for debug purposes)
      operationId: listGroups
      tags: [group]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                description: List of groups
                minItems: 0
                maxItems: 1000
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /groups/{id}/members:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    patch:
      summary: Add members to a group
      description: Add one or more users to an existing group by username.
      operationId: addGroupMembers
      tags: [group]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Usernames to add to the group
        content:
          application/json:
            schema:
              type: object
              description: Request body for adding members to a group
              required: [members]
              properties:
                members:
                  type: array
                  minItems: 1
                  maxItems: 100
                  description: List of usernames to add to the group
                  items:
                    type: string
                    minLength: 3
                    maxLength: 32
                    pattern: '^[a-zA-Z0-9_]+$'
                    description: Username to add to the group
      responses:
        '200':
          description: Members added to the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Leave group
      description: Leave a group by removing the current user from it
      operationId: leaveGroup
      tags: [group]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Left the group
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /groups/{id}/name:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    patch:
      summary: Set group name
      description: Set or update the name of a group
      operationId: setGroupName
      tags: [group]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for updating group name
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating group name
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: '^.*$'
                  description: New group name
      responses:
        '200':
          description: Name changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /groups/{id}/photo:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    patch:
      summary: Set group photo
      description: Set or update the photo of a group
      operationId: setGroupPhoto
      tags: [group]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for updating the group photo
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating the group photo
              required: [photoUrl]
              properties:
                photoUrl:
                  type: string
                  format: uri
                  minLength: 10
                  maxLength: 2048
                  pattern: '^https?://.*$'
                  description: New group photo URL
      responses:
        '200':
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'